requirements.txt
pyopenssl
requests==2.18.4
pytest==2.9.0
PyHamcrest
Faker
pyyaml==3.11
ndg-httpsclient
certifi
